refactor(Core UI): :recycle: Organize charts and custom cell fodlers
