+++ __tests__/SkLinkCell.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import SkLinkCell from '../src/core/components/SkLinkCell';
+import SkLinkCell from '../src/core/components/SkTable/SkCustomCells/SkLinkCell';
 import { Providers } from '../src/providers';
 
 describe('SkLinkCell', () => {
