+++ src/core/components/SkCharts/SkChartArea/SkChartArea.constants.ts
