+++ src/core/components/SkCharts/SkChartPie/SkChartPie.constants.ts
@@ -1,0 +1,8 @@
+const CHART_CONFIG = {
+  LAYOUT: {
+    DEFAULT_PADDING: { bottom: 40, left: 40, right: 40, top: 40 },
+    DEFAULT_HEIGHT: 300
+  }
+};
+
+export { CHART_CONFIG };
