+++ src/core/components/SkTable/SkCustomCells/SkViewDetailsCell/index.tsx
