+++ src/pages/Services/services/index.ts
@@ -1,6 +1,6 @@
 import { DEFAULT_COMPLEX_STRING_SEPARATOR } from '../../../config/app';
 import { styles } from '../../../config/styles';
-import { DEFAULT_SANKEY_CHART_FLOW_VALUE } from '../../../core/components/SKSanckeyChart/SkSankey.constants';
+import { DEFAULT_SANKEY_CHART_FLOW_VALUE } from '../../../core/components/SkCharts/SKSanckeyChart/SkSankey.constants';
 import { removeDuplicatesFromArrayOfObjects } from '../../../core/utils/removeDuplicatesFromArrayOfObjects';
 import { GraphCombo, GraphEdge, GraphElementNames, GraphIconKeys, GraphNode } from '../../../types/Graph.interfaces';
 import { ConnectorResponse, ListenerResponse } from '../../../types/REST.interfaces';
